Tests added to InfoSection, refactored tests into HomePage single test file"
