+++ src/__tests__/App.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { MemoryRouter, useLocation } from 'react-router-dom';
-import App from './components/App';
+import App from '../components/App';
 
 const LocationDisplay = () => {
   const location = useLocation();
